# swagger.yml
swagger: "2.0"
info:
  title: Brigham Young University Audiovisual API
  description: Manage various AV systems across BYU campus with the BYU AV API
  version: "0.4"
# host: api.byu.edu
host: avmetrics1.byu.edu:8000
schemes:
  - http
# basePath: "/v0.3"
produces:
  - application/json
paths:
...
  /rooms:
    get:
      summary: All Rooms
      description: |
        Returns information about all the TEC rooms across campus.
      tags:
        - Rooms
      responses:
        200:
          description: An array of rooms
          schema:
            type: array
            items:
              $ref: "#/definitions/RoomResponse"
        401:
          $ref: "#/responses/401"
        500:
          $ref: "#/responses/500"
        503:
          $ref: "#/responses/503"
        default:
          $ref: "#/responses/default"
    post:
      summary: Make a Room
      description: |
        Make a new room and add it to the Crestron Fusion database.
      parameters:
        - name: body
          description: The new room to created
          in: body
          required: true
          schema:
            $ref: "#/definitions/RoomRequest"
      tags:
        - Rooms
      responses:
        201:
          description: All went well
          schema:
            $ref: "#/definitions/RoomResponse"
        401:
          $ref: "#/responses/401"
        500:
          $ref: "#/responses/500"
        503:
          $ref: "#/responses/503"
        default:
          $ref: "#/responses/default"
  /rooms/{room}:
    get:
      summary: Get a Specific Room
      description: |
        Returns information about the various AV products offered across
        campus. The response includes the display name and other details
        about each product.
      parameters:
        - $ref: "#/parameters/room"
      tags:
        - Rooms
      responses:
        200:
          description: A room object
          schema:
            $ref: "#/definitions/RoomResponse"
        401:
          $ref: "#/responses/401"
        500:
          $ref: "#/responses/500"
        503:
          $ref: "#/responses/503"
        default:
          $ref: "#/responses/default"
    put:
      summary: Update a Room
      description: |
        Update information about a room in the Crestron Fusion database.
      parameters:
        - $ref: "#/parameters/room"
        - name: body
          description: The updated room information
          in: body
          required: true
          schema:
            $ref: "#/definitions/RoomRequest"
      tags:
        - Rooms
      responses:
        200:
          description: All went well
          schema:
            $ref: "#/definitions/RoomResponse"
        401:
          $ref: "#/responses/401"
        500:
          $ref: "#/responses/500"
        503:
          $ref: "#/responses/503"
        default:
          $ref: "#/responses/default"
    delete:
      summary: Delete a Room
      description: |
        Delete a room from the Crestron Fusion database.
      parameters:
        - $ref: "#/parameters/room"
      tags:
        - Rooms
      responses:
        200:
          description: All went well
        401:
          $ref: "#/responses/401"
        500:
          $ref: "#/responses/500"
        503:
          $ref: "#/responses/503"
        default:
          $ref: "#/responses/default"
...
  /buildings/{building}/rooms:
    get:
      summary: Get All Building Rooms
      description: |
        Returns information about all TEC rooms in a specified buliding.
      parameters:
        - $ref: "#/parameters/building"
      tags:
        - Rooms
        - Buildings
      responses:
        200:
          description: An array of rooms
          schema:
            $ref: "#/definitions/RoomResponse"
        401:
          $ref: "#/responses/401"
        500:
          $ref: "#/responses/500"
        503:
          $ref: "#/responses/503"
        default:
          $ref: "#/responses/default"
  /buildings/{building}/rooms/{room}:
    get:
      summary: Get a Specific Room
      description: |
        Returns information about a specific TEC room in a specified buliding.
      parameters:
        - $ref: "#/parameters/building"
        - $ref: "#/parameters/room"
      tags:
        - Rooms
        - Buildings
      responses:
        200:
          description: A room object
          schema:
            $ref: "#/definitions/RoomResponse"
        401:
          $ref: "#/responses/401"
        500:
          $ref: "#/responses/500"
        503:
          $ref: "#/responses/503"
        default:
          $ref: "#/responses/default"
    put:
      summary: Make a Signal
      description: |
        Update information about a room for a specific building and add it to
        the Crestron Fusion database.
      parameters:
        - $ref: "#/parameters/building"
        - $ref: "#/parameters/room"
        - name: body
          description: The updated room information
          in: body
          schema:
            $ref: "#/definitions/RoomRequest"
      tags:
        - Rooms
        - Buildings
      responses:
        201:
          description: All went well
          schema:
            $ref: "#/definitions/RoomResponse"
        401:
          $ref: "#/responses/401"
        500:
          $ref: "#/responses/500"
        503:
          $ref: "#/responses/503"
        default:
          $ref: "#/responses/default"
    delete:
      summary: Delete a Room
      description: |
        Delete a room from the Crestron Fusion database.
      parameters:
        - $ref: "#/parameters/building"
        - $ref: "#/parameters/room"
      tags:
        - Buildings
        - Rooms
      responses:
        200:
          description: All went well
        401:
          $ref: "#/responses/401"
        500:
          $ref: "#/responses/500"
        503:
          $ref: "#/responses/503"
        default:
          $ref: "#/responses/default"
...
definitions:
  HATEOAS:
    type: object
    properties:
      rel:
        type: string
        description: A human-readable name for what the link represents
      href:
        type: string
        description: The link to the endpoint
...
  Room:
    type: object
    properties:
      links:
        type: array
        items:
          $ref: "#/definitions/HATEOAS"
      name:
        type: string
        description: The human-readable name of the room
      ID:
        type: string
        description: The GUID of the room
      roomID:
        type: string
        description: The room number
      description:
        type: string
        description: A human-readable description of the room
      address:
        type: string
        description: The IP address of the processor
      signals:
        type: array
        description: The available signals attached to this processor
        items:
          $ref: "#/definitions/SignalRequest"
      available:
        type: boolean
        description: |
          Whether or not the room is currently available (calculated based on
          campus room scheduling and whether or not the room"s processor
          reports as in-use)
...
  RoomResponse:
    type: object
    properties:
      links:
        type: array
        items:
          $ref: "#/definitions/HATEOAS"
      rooms:
        type: array
        items:
          $ref: "#/definitions/Room"
  RoomRequest:
    type: object
    properties:
      name:
        type: string
        description: The human-readable name of the room
      roomID:
        type: string
        description: The room number
      description:
        type: string
        description: A human-readable description of the room
...
  SignalRequest:
    type: object
    properties:
      name:
        type: string
        description: The human-readable name of the signal
      type:
        type: integer
        description: Crestron-defined signal type
  Error:
    type: object
    properties:
      message:
        type: string
parameters:
  building:
    name: building
    in: path
    description: Building shortname
    required: true
    type: string
  room:
    name: room
    in: path
    description: Room name
    required: true
    type: string
...
responses:
  401:
    description: You are not authorized to make that request
    schema:
      $ref: "#/definitions/Error"
  500:
    description: There was a problem with the server on our end
    schema:
      $ref: "#/definitions/Error"
  503:
    description: The server is overloaded, is under maintenance, or is otherwise unavailable
    schema:
      $ref: "#/definitions/Error"
  default:
    description: Unexpected error
    schema:
      $ref: "#/definitions/Error"